# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoalMind API
  version: 1.0.0
  description: |
    GoalMind backend contract for the diploma project frontend.

    Success responses use a consistent envelope:
      { "data": ... }

    Error responses use:
      {
        "error": {
          "code": "string",
          "message": "string",
          "details": {}
        }
      }
servers:
  - url: http://localhost:8080
    description: Local development backend
...
security:
  - bearerAuth: []
paths:
  /api/requests/analyze:
    post:
      tags: [Requests]
      summary: Analyze a freeform user request
      operationId: analyzeRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeFreeformRequest'
            examples:
              default:
                value:
                  text: изучить Golang за 3 месяца
                  provider: gemini
      responses:
        '200':
          description: Request analyzed and structured clarification questions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeFreeformEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/requests/{requestId}/answers:
    post:
      tags: [Requests]
      summary: Save structured answers for a previously analyzed request
      operationId: submitRequestAnswers
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRequestAnswers'
            examples:
              default:
                value:
                  answers:
                    - questionId: current_level
                      value: beginner
                    - questionId: available_hours_per_day
                      value: 3
                    - questionId: focus_topics
                      value: [practice, project]
      responses:
        '200':
          description: Answers saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitRequestAnswersEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/requests/{requestId}/evaluate:
    post:
      tags: [Requests]
      summary: Evaluate request feasibility after structured answers
      operationId: evaluateRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Feasibility evaluation returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateRequestEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/requests/{requestId}/generate-plan:
    post:
      tags: [Requests]
      summary: Generate a Goal and TaskTree from a freeform request and saved answers
      operationId: generatePlanFromRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '201':
          description: Goal and task tree created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratePlanEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/ai-settings:
    get:
      tags: [Requests]
      summary: Get current user's AI provider settings without exposing API keys
      operationId: getAISettings
      responses:
        '200':
          description: AI settings returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISettingsEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Requests]
      summary: Save current user's AI provider API key
      operationId: updateAISettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AISettingsRequest'
            examples:
              gemini:
                value:
                  provider: gemini
                  apiKey: AIza...
      responses:
        '200':
          description: AI settings saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISettingsEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/oauth/google/start:
    get:
      tags: [Auth]
      summary: Start Google OAuth authorization flow
      operationId: startGoogleOAuth
      security: []
      responses:
        '302':
          description: Redirect to Google OAuth consent screen
          headers:
            Location:
              description: Google OAuth consent screen URL
              schema:
                type: string
                format: uri
              example: https://accounts.google.com/o/oauth2/v2/auth?client_id=demo-client
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/oauth/google/callback:
    get:
      tags: [Auth]
      summary: Handle Google OAuth callback and redirect to frontend
      operationId: handleGoogleOAuthCallback
      security: []
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: OAuth authorization code from Google
        - in: query
          name: state
          required: true
          schema:
            type: string
          description: CSRF protection state token
      responses:
        '302':
          description: Redirect to frontend callback route with JWT token
          headers:
            Location:
              description: Frontend callback URL containing the issued JWT
              schema:
                type: string
                format: uri
              example: http://localhost:5173/auth/callback?token=jwt.token.value
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/goals:
    get:
      tags: [Goals]
      summary: List goals for the authenticated user
      operationId: listGoals
      responses:
        '200':
          description: Goals returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalsEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Goals]
      summary: Create a new goal
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
            examples:
              default:
                value:
                  title: Launch GoalMind diploma MVP
                  description: Prepare a polished demo version for thesis defense
                  category: Education Product
                  priority: HIGH
                  deadline: 2026-06-15
                  availableHoursPerWeek: 18
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/goals/{goalId}:
    get:
      tags: [Goals]
      summary: Get goal details
      operationId: getGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Goals]
      summary: Update a goal
      operationId: updateGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Goal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Goals]
      summary: Delete a goal
      operationId: deleteGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Goal deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/goals/{goalId}/context:
    get:
      tags: [Goal Context]
      summary: Get goal context
      operationId: getGoalContext
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal context returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalContextEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Goal Context]
      summary: Create goal context
      operationId: createGoalContext
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertGoalContextRequest'
      responses:
        '201':
          description: Goal context created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalContextEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Goal Context]
      summary: Update goal context
      operationId: updateGoalContext
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertGoalContextRequest'
      responses:
        '200':
          description: Goal context updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalContextEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/goals/{goalId}/tasks:
    get:
      tags: [Tasks]
      summary: Get the task tree for a goal
      operationId: listGoalTasks
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal tasks returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskTreeNodesEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Tasks]
      summary: Create a task for a goal
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/tasks/{taskId}/status:
    patch:
      tags: [Tasks]
      summary: Update task status
      operationId: updateTaskStatus
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatusRequest'
      responses:
        '200':
          description: Task status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/tasks/{taskId}/dependencies/{dependencyId}:
    delete:
      tags: [Dependencies]
      summary: Delete a task dependency
      operationId: deleteTaskDependency
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/DependencyId'
      responses:
        '204':
          description: Dependency deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequestId:
      in: path
      name: requestId
      required: true
      description: Freeform planning request identifier
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_error
              message: One or more request fields are invalid.
              details:
                field: email
    Unauthorized:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Authentication token is missing or expired.
              details: {}
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: internal_error
              message: An unexpected error occurred.
              details: {}
  schemas:
...
    AnalyzeFreeformRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        provider:
          type: string
          enum: [gemini, deepseek]
          description: Optional provider override. If omitted, user settings or server default are used.
      example:
        text: изучить Golang за 3 месяца
        provider: gemini
...
    AnalyzeFreeformEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AnalyzeFreeformResponse'
...